Small cosmetic improvements
Added a third histogram with "credible"  news to contrast with not
credible one. Reduced height of histogram rows. Only the top one has
zoom control. Only the bottom one has x-axis labels.

Added googlefonts which are more in-line with the project's brand
identity.

Dashboards can now specify a "version" string, which is displayed next
to the title, but in smaller font.

Tweaked colors in histogram.