Custom doc+review components in table. Added Sidebar.
The table panel now has a custom default view called "Doc+Review"
which displays the document and its credibility review as "cards". The
document type is used so that tweets are displayed as embedded tweets
using the official Twitter widget. For documents which are not tweet,
we have a custom card.

Credibility reviews (or at least the main values encoded in documents)
returned by the back-end are also displayed as a card.

Banana does not support sidebars by default. This is problematic
because filters which affect the whole dashboard would be much better
displayed in such a sidebar. This commit implements an experimental
sidebar. The idea is that it's similar to a row in the normal
dashboard; however at the moment drag-and-drop doesn't work so panels
need to be added "programmatically" (by editing the json config
manually).

Fixed bug in heatmap: when clicking on a cell when axis are
transposed, the resulting filters were swapped.

Added footer and improvements to the standard co-inform css theme.