Continued refactor, simplify and document of reviewGraph
Started refactoring reviewGraphDir. I noticed the first thing it did
was remove orphan and irrelevant nodes/links. This is conceptually
still pre-processing of the acred graph, so I moved this to the
rgProcessor.

There were also several graph search methods, which I moved to
rgProcessor.Searcher.

Finally, reviewGraphDir was full of Object.getPrototypeOf calls. This
was because, it wraps the nodes and links it receives to Objects, and
since the rgProcessor adds the state/calculated values to the original
graph, these are not available directly in the nodes/links the
reviewGraphDir is seeing. It makes sense to separate the original
graph's nodes and links while all additional properties needed for
displaying the reviewGraph are added at the new Objects level. This
means that the rgProcessor should add all these new properties to the
new Objects, which also means it should be responsible for creating
the new Objects in the first place. This, however, meant that some of
the search functionalities had to be revised/reimplemented.

One of the drawbacks of the current acred review graph is that a
single node can have multiple uniquely identifying strings. So instead
of only relying on property id, I generalised this so that we can look
for the node based on all the available identifying values. Also, for
performance and separation of concerns, the Searcher creates an index
of node identifiers to the nodes.