Start moving events from directive to controller
Started moving events and handlers from reviewGraph directive to the
reviewGraph controller. The former should only deal with the d3 force
simulation (and events occurring with the SVG nodes). The latter
should coordinate interactions between the force simulation and other
components like the cards.

For now moved mainly events related to selecting a node in the graph.